reorg integration tests to allow parallel run
