+++ stats/stats-server/tests/it/common.rs
@@ -1,0 +1,15 @@
+use reqwest::{RequestBuilder, Response};
+
+pub async fn send_arbitrary_request(request: RequestBuilder) -> Response {
+    let response = request
+        .send()
+        .await
+        .unwrap_or_else(|_| panic!("Failed to send request"));
+
+    if !response.status().is_success() {
+        let status = response.status();
+        let message = response.text().await.expect("Read body as text");
+        panic!("Invalid status code (success expected). Status: {status}. Message: {message}")
+    }
+    response
+}
