+++ stats/stats-server/tests/it/main.rs
@@ -1,0 +1,5 @@
+mod common;
+
+mod counters;
+mod lines;
+mod swagger;
